Cleanup Overview react UI code
- refactor into container pattern for main container logic + views
- split up views a bit more
- fix up proptypes
- cleanup container methods
