+++ src/overview/components/Header.js
@@ -1,0 +1,48 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+
+import DateRangeSelection from './DateRangeSelection'
+import styles from './Overview.css'
+
+const Header = ({ currentQueryParams: { query, startDate, endDate }, ...props }) => (
+    <div className={styles.navbar}>
+        <div className={styles.logo} />
+        <div className={styles.searchField}>
+            <input
+                className={styles.query}
+                onChange={props.onInputChange}
+                placeholder='Search your memory'
+                value={query}
+                ref={props.setInputRef}
+            />
+            <DateRangeSelection
+                startDate={startDate}
+                endDate={endDate}
+                onStartDateChange={props.onStartDateChange}
+                onEndDateChange={props.onEndDateChange}
+            />
+        </div>
+        <div className={styles.links}>
+            <a href='/options/options.html'>
+                <img
+                    src='/img/settings-icon.png'
+                    className={styles.icon}
+                />
+            </a>
+        </div>
+    </div>
+)
+
+Header.propTypes = {
+    currentQueryParams: PropTypes.shape({
+        query: PropTypes.string,
+        startDate: PropTypes.number,
+        endDate: PropTypes.number,
+    }).isRequired,
+    setInputRef: PropTypes.func.isRequired,
+    onInputChange: PropTypes.func.isRequired,
+    onStartDateChange: PropTypes.func.isRequired,
+    onEndDateChange: PropTypes.func.isRequired,
+}
+
+export default Header
