+++ src/overview/index.js
@@ -2,6 +2,6 @@ import reducer from './reducer'
 import * as actions from './actions'
 import * as epics from './epics'
 import enhancer from './enhancer'
-import Overview from './components/Overview'
+import container from './container'
 
-export default {reducer, actions, epics, enhancer, components: { Overview }}
+export default { reducer, actions, epics, enhancer, container }
